+++ game/src/objects/resource/ResourceFactory.h
@@ -15,11 +15,11 @@ public:
 	ResourceFactory() = default;
 	~ResourceFactory() = default;
 
-	ResourceEntity* create(int id, Urho3D::IntVector2 bucketCords);
+	ResourceEntity* create(unsigned short id, Urho3D::IntVector2 bucketCords);
 	ResourceEntity* load(dbload_resource* resource) const;
 	void setResUid(unsigned resUid);
 
 private:
-	ResourceEntity* create(int id, Urho3D::IntVector2 bucketCords, UId uid) const;
+	ResourceEntity* create(unsigned short id, Urho3D::IntVector2 bucketCords, UId uid) const;
 	unsigned currentUId;
 };
